Update README.md
Metallurgical industries want to be able to effectively control the quality of their products in order to maintain their prestige and increase their customers base. Therefore, the design and implementation of an optimal  technical scheme to detect these metal defects would be a solution to this problem. In this work, two methods used to extract local characteristics from the images of metal products are presented. These are the Local Binary Pattern (LBP), the Grey Level Co-occurrence Matrix (GLCM). After extracting the specific characteristics describing each image, these informations are grouped into a matrix on which the spectral clustering technique will be applied in order to classify images according to their similarities. Each of these algorithms is applied to the tabulated images. The effectiveness of these methods is evaluated in terms of the number of positive detections obtained.The results have shown that the combination of Local Binary Pattern and spectral clustering gives a better detection accuracy compared to the Grey Level Co-occurrence Matrix.

Data available  are images of tabulators. These images have been divided into three basic classes. The first class is that of tabulators assumed to be well manufactured (Good tabulator), the second class is that of tabulators for which the component has been omitted during the manufacturing process (Empty tabulator) and the third class is that of tabulators poorly made during the manufacturing process (Bad tabulator). The distinction between good and bad tabulators has been made by experts on the plan.